Implement safer texture binding
In order to bind a texture you need a texture target.
To aqcuire a texture target you need to activate a texture unit.
To activate a texture unit you need a texture unit slot.

This concept has to be extended to all things that can be "bound" to
minimize mistakes that can be made due to OpenGL calls being stateful.

Also need to define some enums and implement some more functions but
will do that on a when needed basis.